New Python script called extract_VC_strain_names_from_filenames.py extracts the strain names from a .txt file called 'MMP_combined_vcf_filenames.txt' generated with the shell command: find . name "VC*.combined.vcf" | sort | uniq > MMP_combined_vcf_filenames.txt which is a list of .vcf filenames containing strain names that starts with VC and are < 8 characters long and writes the output to a .txt file called list_VCstrains_vcf.txt
The purpose of generating a file containing a list of all the 2007 MMP strains is so that I can use Stephane's
perl script (create_gwas_vcf_MMP.pl) to generate a large merged .vcf file for every MMP strain (all 2007). This
will be useful to test Stephane's new perl script to subset .vcf files for a given list of MMP strains from the
large merged .vcf file containing all the strains. It will also be useful to do SKAT on Nigel's MMP data.